Move music config to new package
