+++ src/main/java/com/raikuman/troubleclub/radio/commands/music/Volume.java
@@ -3,7 +3,7 @@ package com.raikuman.troubleclub.radio.commands.music;
 import com.raikuman.botutilities.commands.manager.CategoryInterface;
 import com.raikuman.botutilities.helpers.RandomColor;
 import com.raikuman.troubleclub.radio.category.MusicCategory;
-import com.raikuman.troubleclub.radio.config.MusicDB;
+import com.raikuman.troubleclub.radio.config.music.MusicDB;
 import com.raikuman.troubleclub.radio.music.GuildMusicManager;
 import com.raikuman.troubleclub.radio.music.PlayerManager;
 import com.raikuman.botutilities.commands.manager.CommandContext;
@@ -20,7 +20,7 @@ import java.util.List;
 /**
  * Handles setting the volume of the bot
  *
- * @version 1.6 2022-13-07
+ * @version 1.7 2022-15-07
  * @since 1.1
  */
 public class Volume implements CommandInterface {
